% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariancecorrelation.R
\name{Correlation}
\alias{Correlation}
\title{\code{Correlation}}
\usage{
Correlation(x, y, weights = NULL, correlation = TRUE)
}
\arguments{
\item{x}{A numeric vector.}

\item{y}{A numeric vector.}

\item{weights}{A numeric vector containing the value of the weight for each
row of \code{data}. If weights is NULL then this function is just a wrapper
for the base functions \code{cov} and \code{cor}.}

\item{correlation}{If \code{FALSE}, returns the covariance.}
}
\description{
Computes the correlation between two vectors, using weights if provided.
}
\examples{
x <- 1:10
y <- c(2, 2:10)
w <- rep(1, 10)
Correlation(x, y, w)
}
