% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariancecorrelation.R
\name{CorrelationMatrix}
\alias{CorrelationMatrix}
\title{\code{CorrealtionMatrix}}
\usage{
CorrelationMatrix(input.type = "Variables", input.data, use.names = FALSE,
  ignore.columns = "", missing.data = "Use partial data",
  spearman = FALSE, filter = NULL, weights = NULL)
}
\arguments{
\item{input.type}{TODO}

\item{input.data}{TODO}

\item{use.names}{TODO}

\item{ignore.columns}{TODO}

\item{missing.data}{TODO}

\item{spearman}{TODO}

\item{filter}{TODO}

\item{weights}{An optional vector of sampling weights.}
}
\description{
Produces a correlation TODO
}
