% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosinesimilarity.R
\name{CosineSimilarities}
\alias{CosineSimilarities}
\title{\code{CosineSimilarities}}
\usage{
CosineSimilarities(x, weight = NULL)
}
\arguments{
\item{x}{A matrix whose columns we use to calculate the cosine similarity.}

\item{weight}{A vector of weights.}
}
\description{
Computes the cosine similarity between columns of a matrix. The cosine similarity
is the cosine of the angle between vectors (columns).
}

