% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency.R
\name{Frequency}
\alias{Frequency}
\title{\code{Frequency}}
\usage{
Frequency(x, subset = NULL, weights = NULL)
}
\arguments{
\item{x}{The variable to be tablulated.}

\item{subset}{An optional subset.}

\item{weights}{an optional weight vector.}
}
\description{
Computes a frequency table, taking weights into account if applicable.
}
