% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiway.R
\name{Interaction}
\alias{Interaction}
\title{Interaction}
\usage{
Interaction(data, subset)
}
\arguments{
\item{data}{A \code{\link{data.frame}} of variables.}

\item{subset}{The sub-group to include in the analysis.}
}
\description{
Interaction
}
