% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_and_min.R
\name{Max}
\alias{Max}
\title{Max}
\usage{
Max(x, weights = NULL)
}
\arguments{
\item{x}{A \code{\link{data.frame}} or  \code{\link{matrix}}.}

\item{weights}{The sampling or replication weights.}
}
\description{
Computes the weighted maximum of one or more variables. Missing values are automatically excluded.
}
\details{
Computes the maximum of the observations with positive weights.
}
