% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.R
\name{Mean}
\alias{Mean}
\title{Mean}
\usage{
Mean(x, weights = NULL)
}
\arguments{
\item{x}{A \code{\link{data.frame}} or  \code{\link{matrix}}.}

\item{weights}{The sampling or replication weights.}
}
\description{
Computes the weighted mean of one or more variables. Missing values are automatically excluded.
}

