% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumofsquares.R
\name{OmegaSquared}
\alias{OmegaSquared}
\title{OmegaSquared}
\usage{
OmegaSquared(x, group, weights = NULL)
}
\arguments{
\item{x}{A \code{\link{data.frame}} or  \code{\link{matrix}}.}

\item{group}{A variable indicating group membership. Either a factor or coerced to a factor.}

\item{weights}{The sampling or replication weights.}
}
\description{
The generalized R-Squared (i.e., with multiple variables).
}

