% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplotmatrix.R
\name{ScatterplotMatrix}
\alias{ScatterplotMatrix}
\title{Scatterplot matrix}
\usage{
ScatterplotMatrix(
  x,
  seed = 123,
  symbol = "Small dots",
  modifications = "None",
  fit.type = "None",
  weights = 1:NROW(x)
)
}
\arguments{
\item{x}{A dataframe containing the variables to be plotted.}

\item{seed}{A seed used to generate random variables for jitter.}

\item{symbol}{One of "Small dots", "Dots", "Big dots", "Circles".}

\item{modifications}{One of "None", "Jitter", or "Enlarge points with multiple observations".}

\item{fit.type}{One of "None", "LOESS", or "Straight".}

\item{weights}{A numeric vector with length equal to the number of rows in \code{x}.}
}
\description{
Creates a matrix of scatterplot from variable pairs
}
