% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariancecorrelation.R
\name{SpearmanRanks}
\alias{SpearmanRanks}
\title{\code{SpearmanRanks}}
\usage{
SpearmanRanks(x, weights)
}
\arguments{
\item{x}{A numeric vector.}

\item{weights}{A vector of weights. Must not have missing values.}
}
\description{
Computes ranks for computing Spearman correlation.
}
