% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standarddeviation.R
\name{StandardDeviation}
\alias{StandardDeviation}
\title{StandardDeviation}
\usage{
StandardDeviation(x, weights = NULL, method = "SPSS")
}
\arguments{
\item{x}{A \code{\link{data.frame}} or  \code{\link{matrix}}.}

\item{weights}{The sampling or replication weights.}

\item{method}{The way that the weights are addressed in the computation. Currently only "SPSS" is supported.}
}
\description{
Computes the weighted variance of one or more variables
}

