% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statisticbygroup.R
\name{StatisticsByGroup}
\alias{StatisticsByGroup}
\title{StatisticsByGroup}
\usage{
StatisticsByGroup(x, group, weights = NULL, FUN = Mean)
}
\arguments{
\item{x}{A \code{\link{data.frame}} or  \code{\link{matrix}}.}

\item{group}{A variable indicating group membership. Either a factor or coerced to a factor.}

\item{weights}{The sampling or replication weights.}

\item{FUN}{A function that can handle a weight argument.}
}
\description{
StatisticsByGroup
}
\examples{
## Two variables with a single observation and a single group
dat <- matrix(c(1.0, 2.0), 1, 2)
stopifnot(identical(unname(StatisticsByGroup(dat, 1)), dat))
}
