% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum.R
\name{Sum}
\alias{Sum}
\title{Sum}
\usage{
Sum(x, weights = NULL)
}
\arguments{
\item{x}{A \code{\link{data.frame}} or  \code{\link{matrix}}.}

\item{weights}{The sampling or replication weights.}
}
\description{
Computes the weighted sum  of one or more variables. Missing values are automatically excluded.
}
