% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumofsquares.R
\name{SumOfSquares}
\alias{SumOfSquares}
\title{SumOfSquares}
\usage{
SumOfSquares(x, weights = NULL)
}
\arguments{
\item{x}{A \code{\link{data.frame}} or  \code{\link{matrix}}.}

\item{weights}{The sampling or replication weights.}
}
\description{
Computes the weighted sum of squares of one or more variables
}

