% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumofsquares.R
\name{SumOfSquaresByGroup}
\alias{SumOfSquaresByGroup}
\title{SumOfSquares}
\usage{
SumOfSquaresByGroup(x, group, weights = NULL)
}
\arguments{
\item{x}{A \code{\link{data.frame}} or  \code{\link{matrix}}.}

\item{group}{A variable indicating group membership. Either a factor or coerced to a factor.}

\item{weights}{The sampling or replication weights.}
}
\description{
Computes the weighted sum of squares of one or more variables
}
