% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumofsquares.R
\name{TotalSumOfSquares}
\alias{TotalSumOfSquares}
\title{TotalSumOfSquares}
\usage{
TotalSumOfSquares(x, weights = NULL)
}
\arguments{
\item{x}{A \code{\link{data.frame}} or  \code{\link{matrix}}.}

\item{weights}{The sampling or replication weights.}
}
\description{
Computes the tss for one or more variables
}

