% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{WeightedCounts}
\alias{WeightedCounts}
\title{\code{WeightedCounts} Computes a SVD with frequency weights.}
\usage{
WeightedCounts(x, weights)
}
\arguments{
\item{x}{A factor.}

\item{weights}{Frequency Weights.}
}
\description{
\code{WeightedCounts} Computes a SVD with frequency weights.
}
