% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariancecorrelation.R
\name{CorrelationsWithSignificance}
\alias{CorrelationsWithSignificance}
\title{\code{CorrelationsWithSignificance}}
\usage{
CorrelationsWithSignificance(data, weights, spearman = FALSE)
}
\arguments{
\item{data}{A numeric data frame}

\item{weights}{A vector of weights}

\item{spearman}{Whether to compute Spearman correlations instead of Pearson correlations.}
}
\description{
Computes correlations with significance from zero.
}

