% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiway.R
\name{Multiway}
\alias{Multiway}
\title{\code{Multiway}}
\usage{
Multiway(
  rows,
  columns = NULL,
  numeric = NULL,
  numeric.statistic = "Mean",
  hide.empty.rows = FALSE,
  hide.empty.columns = FALSE,
  subset = NULL,
  weights = NULL
)
}
\arguments{
\item{rows}{A \code{\link{data.frame}} of variables to show in the rows of the table.}

\item{columns}{An optional \code{\link{data.frame}} of variables to show in the columns of the table.}

\item{numeric}{An optional numeric variable, which is used to compute the mean in the cells of the table.}

\item{numeric.statistic}{One of \code{"Mean"}, \code{"Minimum"}, \code{"Maximum"}, \code{"Sum"}.The statistic to be computed when rows and columns have been specified. Ignored if
no columns have been specified.}

\item{hide.empty.rows}{Hide rows containing no data.}

\item{hide.empty.columns}{Hide columns containing no data.}

\item{subset}{The sub-group to include in the analysis.}

\item{weights}{The sampling or replication weights.}
}
\description{
Creates a crosstab by aggregating numeric data over factors.
}
