% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariancecorrelation.R
\name{print.CorrelationMatrix}
\alias{print.CorrelationMatrix}
\title{\code{print.CorrelationMatrix}}
\usage{
\method{print}{CorrelationMatrix}(x, ..., show.cell.values = "Automatic",
  row.labels = "Yes", column.labels = "Yes")
}
\arguments{
\item{x}{An object of class \code{\link{CorrelationMatrix}}.}

\item{...}{TODO}

\item{show.cell.values}{TODO}

\item{row.labels}{TODO}

\item{column.labels}{TODO}
}
\description{
\code{print.CorrelationMatrix}
}
\details{
Displays a correlation matrix as a heatmap.
}
