% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariancecorrelation.R
\name{print.CorrelationMatrix}
\alias{print.CorrelationMatrix}
\title{\code{print.CorrelationMatrix}}
\usage{
\method{print}{CorrelationMatrix}(x, ...)
}
\arguments{
\item{x}{An object of class \code{\link{CorrelationMatrix}}.}

\item{...}{Other paramaters, not used.}
}
\description{
\code{print.CorrelationMatrix}
}
\details{
Displays a correlation matrix as a heatmap.
}
