% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statisticbygroup.R
\name{StatisticsByGroup}
\alias{StatisticsByGroup}
\title{StatisticsByGroup}
\usage{
StatisticsByGroup(x, group, weights = NULL, FUN = mean)
}
\arguments{
\item{x}{A \code{\link{data.frame}} or  \code{\link{matrix}}.}

\item{group}{A variable indicating group membership. Either a factor or coerced to a factor.}

\item{weights}{The sampling or replication weights.}
}
\description{
StatisticsByGroup
}

